Config enzyme of jest
