+++ package.json
@@ -48,6 +48,7 @@
     "css-loader": "^0.28.7",
     "element-theme-default": "^1.4.8",
     "enzyme": "^3.1.0",
+    "enzyme-adapter-react-16": "^1.0.2",
     "eslint": "^3.19.0",
     "eslint-config-airbnb": "^15.0.1",
     "eslint-plugin-flowtype": "^2.34.1",
@@ -66,7 +67,6 @@
     "raw-loader": "^0.5.1",
     "react-hot-loader": "^3.1.1",
     "react-scroll-up": "^1.3.3",
-    "react-test-renderer": "^16.0.0",
     "regenerator-runtime": "^0.11.0",
     "sass-loader": "^6.0.5",
     "sinon": "^2.4.1",
@@ -92,8 +92,7 @@
   "jest": {
     "testRegex": "(/jest/.+\\.(js|jsx)$)|_test.jsx?$",
     "testPathIgnorePatterns": [
-      "<rootDir>/tests/jest/config.js",
-      "<rootDir>/tests/jest/mock.js",
+      "<rootDir>/tests",
       "<rootDir>/dist"
     ],
     "moduleNameMapper": {
