+++ src/rate/__test__/rate_test.jsx
@@ -1,5 +1,4 @@
 import React from 'react';
-import renderer from 'react-test-renderer';
 import { mount } from 'enzyme';
 
 import { Rate } from '../../../src';
