+++ tests/jest/config.js
@@ -1,3 +1,5 @@
+require('./polyfill');
+
 const Enzyme = require('enzyme');
 const EnzymeAdapter = require('enzyme-adapter-react-16');
