+++ tests/jest/polyfill.js
@@ -1,0 +1,4 @@
+
+global.requestAnimationFrame = (cb) => {
+  setTimeout(cb, 0)
+}
